CPS-393: Refactor Gulp Task
